requirements.txt
requests
beautifulsoup4
pandas
openpyxl
gspread
google-auth
schedule